Module type Dirsp_proscript_mirage.CUSTOM_OPTIONS

val random_bytes : int -> string -> Stdlib.Bytes.t

random_bytes sz id creates sz random bytes with a hint id.

If and only if you have a unit test and need determinism, use the id to deterministically create random numbers.

Every other implementation MUST ignore the id and generate a truly random sequence.